comb through all methods and list them in spec file
I like this practice of going through all methods in the main method and listing them, classifying them.
I haven't included the decipher method because it was not part of the assignment and if I can do the testing on the rest of the methods, I'm sure I can do the testing on decipher too.